requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==25.1.0
    # via univr-chatbot (pyproject.toml)
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
distro==1.9.0
    # via google-genai
fastapi==0.128.0
    # via univr-chatbot (pyproject.toml)
google-auth==2.47.0
    # via google-genai
google-genai==1.57.0
    # via univr-chatbot (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   univr-chatbot (pyproject.toml)
    #   google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via univr-chatbot (pyproject.toml)
markupsafe==3.0.3
    # via jinja2
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   fastapi
    #   google-genai
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.1
    # via
    #   univr-chatbot (pyproject.toml)
    #   uvicorn
python-multipart==0.0.21
    # via univr-chatbot (pyproject.toml)
pyyaml==6.0.3
    # via uvicorn
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via google-genai
starlette==0.50.0
    # via fastapi
tenacity==9.1.2
    # via google-genai
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
uvicorn==0.40.0
    # via univr-chatbot (pyproject.toml)
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
